D2SignupHelper/pyproject.toml
[tool.poetry]
name = "d2signuphelper"
version = "0.1.0"
description = ""
authors = ["Mitchell Frazer <108438939+moosieth@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.7.2"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"

[tool.poetry.scripts]
bot-start = "d2signuphelper.bot:bot_init"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
